requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
lxml==5.3.0
outcome==1.3.0.post0
pandas==2.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
websocket-client==1.8.0
wsproto==1.2.0
webdriver-manager==4.0.2
mysql-connector-python==9.1.0
fastapi==0.115.5
uvicorn==0.32.0
fastapi-mail==1.4.1
pydantic==2.9.2
APScheduler==3.10.4
tenacity==8.5.0
pydantic-settings==2.6.1
